fix comments

+++ extensions-contrib/distinctcount/src/main/java/io/druid/query/aggregation/distinctcount/DistinctCountBufferAggregator.java
@@ -88,9 +88,8 @@ public class DistinctCountBufferAggregator implements BufferAggregator
   @Override
   public double getDouble(ByteBuffer buf, int position)
   {
-    return (double) buf.getLong(position );
+    return buf.getDouble(position);
   }
-
 
   @Override
   public void close()
